Extract logic to an orchestration handler
